+++ src/api/services/index.ts
@@ -4,6 +4,7 @@
 */
 
 export * from "./environment"
+export * from "./feature"
 export * from "./prediction"
 export * from "./predictor"
 export * from "./storage"